Code clean up.

+++ app/src/main/java/com/flowercentral/flowercentralbusiness/util/Logger.java
@@ -11,22 +11,17 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Calendar;
 
-/**
- * Created by Ashish Upadhyay on 2/8/2016.
- */
 public class Logger {
+    private static boolean print_log_to_console = true;
+    public static boolean print_log_to_file = false;
+
     /**
      * used to log an information to  LogCat.
      *
      * @param className : Name of the class file
      * @param methodName : Name of the method
      * @param info : Information to be print in LogCat.
-     * @return Void
      */
-
-    public static boolean print_log_to_console = true;
-    public static boolean print_log_to_file = false;
-
     public static void log(String className, String methodName, String info, int logLevel) {
         if (print_log_to_console) {
             String log = "Method: " + methodName + " | ";
@@ -57,9 +52,8 @@ public class Logger {
      * @param className  : Name of the class file
      * @param methodName : Name of the method
      * @param errorMsg   : Error Message
-     * @return Void
      */
-    public static void logError(String className, String methodName, String errorMsg) {
+    private static void logError(String className, String methodName, String errorMsg) {
         String log = "Class: " + className + " | ";
         log = log + "Method: " + methodName + " | ";
         log = log + "Error: " + errorMsg;
@@ -80,38 +74,44 @@ public class Logger {
             String folderPath = Environment.getExternalStorageDirectory().getPath() + "/" + AppConstant.APP_NAME + "/";
             String filePath = folderPath + "log.txt";
             File directory = new File(folderPath);
-            directory.mkdirs();
-            File logFile = new File(filePath);
-            if (!logFile.exists()) {
+            boolean isDirCreated = directory.mkdirs();
+            if(isDirCreated) {
+                File logFile = new File(filePath);
+                if (!logFile.exists()) {
+                    try {
+                        boolean isFileCreated = logFile.createNewFile();
+                        if(isFileCreated) {
+                            BufferedWriter buf = new BufferedWriter(new FileWriter(logFile, true));
+                            buf.append(AppConstant.APP_NAME);
+                            buf.newLine();
+                            buf.close();
+                        }
+
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+
+                //fetch system date and time and append it to the error text
+                final Calendar c = Calendar.getInstance();
+                int mYear = c.get(Calendar.YEAR);
+                int mMonth = c.get(Calendar.MONTH);
+                int mDay = c.get(Calendar.DAY_OF_MONTH);
+                int mHour = c.get(Calendar.HOUR_OF_DAY);
+                int mMinute = c.get(Calendar.MINUTE);
+                int mSecond = c.get(Calendar.SECOND);
+                String text = mDay + "/" + mMonth + "/" + mYear + " - " + mHour + " : " + mMinute + " : " + mSecond + " :: " + message;
                 try {
-                    logFile.createNewFile();
+                    //BufferedWriter for performance, true to set append to file flag
                     BufferedWriter buf = new BufferedWriter(new FileWriter(logFile, true));
-                    buf.append(AppConstant.APP_NAME);
+                    buf.append(text);
                     buf.newLine();
                     buf.close();
-
                 } catch (IOException e) {
                     e.printStackTrace();
                 }
-            }
-
-            //fetch system date and time and append it to the error text
-            final Calendar c = Calendar.getInstance();
-            int mYear = c.get(Calendar.YEAR);
-            int mMonth = c.get(Calendar.MONTH);
-            int mDay = c.get(Calendar.DAY_OF_MONTH);
-            int mHour = c.get(Calendar.HOUR_OF_DAY);
-            int mMinute = c.get(Calendar.MINUTE);
-            int mSecond = c.get(Calendar.SECOND);
-            String text = mDay + "/" + mMonth + "/" + mYear + " - " + mHour + " : " + mMinute + " : " + mSecond + " :: " + message;
-            try {
-                //BufferedWriter for performance, true to set append to file flag
-                BufferedWriter buf = new BufferedWriter(new FileWriter(logFile, true));
-                buf.append(text);
-                buf.newLine();
-                buf.close();
-            } catch (IOException e) {
-                e.printStackTrace();
+            } else {
+                logError(Logger.class.getSimpleName(), "print to file", "Dir creation failed");
             }
         }
     }
